# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Varese Life API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List all sources
      responses:
        "200":
          description: Sources list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceWithUrls"
    post:
      operationId: createSource
      tags: [sources]
      summary: Create source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceWithUrls"
  /sources/{id}:
    get:
      operationId: getSource
      tags: [sources]
      summary: Get source by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceWithUrls"
    put:
      operationId: updateSource
      tags: [sources]
      summary: Update source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceWithUrls"
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /sources/{id}/scrape:
    post:
      operationId: scrapeSource
      tags: [sources]
      summary: Trigger scrape for a source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scrape triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeResponse"
  /sources/{id}/import:
    post:
      operationId: importEvents
      tags: [sources]
      summary: Import events directly (webhook / n8n / Make)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportEventsRequest"
      responses:
        "200":
          description: Events imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTemplateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /templates/{id}/set-default:
    post:
      operationId: setDefaultTemplate
      tags: [templates]
      summary: Set template as default
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Set as default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USERS
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
  /settings/change-password:
    post:
      operationId: changePassword
      tags: [settings]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # DASHBOARD
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SourceUrl:
      type: object
      properties:
        id:
          type: integer
        sourceId:
          type: integer
        url:
          type: string
        label:
          type: string
          nullable: true
        browseractWorkflowId:
          type: string
          nullable: true
        active:
          type: boolean
      required: [id, sourceId, url, active]
    Source:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        browseractWorkflowId:
          type: string
          nullable: true
        promptOverride:
          type: string
          nullable: true
        active:
          type: boolean
        lastScrapedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        preferredScraper:
          type: string
      required: [id, name, active, createdAt, preferredScraper]
    SourceWithUrls:
      allOf:
        - $ref: "#/components/schemas/Source"
        - type: object
          properties:
            urls:
              type: array
              items:
                $ref: "#/components/schemas/SourceUrl"
          required: [urls]
    CreateSourceRequest:
      type: object
      properties:
        name:
          type: string
        browseractWorkflowId:
          type: string
          nullable: true
        promptOverride:
          type: string
          nullable: true
        active:
          type: boolean
        preferredScraper:
          type: string
        urls:
          type: array
          items:
            $ref: "#/components/schemas/CreateSourceUrlRequest"
      required: [name, urls]
    CreateSourceUrlRequest:
      type: object
      properties:
        url:
          type: string
        label:
          type: string
          nullable: true
        browseractWorkflowId:
          type: string
          nullable: true
        active:
          type: boolean
      required: [url]
    UpdateSourceRequest:
      type: object
      properties:
        name:
          type: string
        browseractWorkflowId:
          type: string
          nullable: true
        promptOverride:
          type: string
          nullable: true
        active:
          type: boolean
        preferredScraper:
          type: string
        urls:
          type: array
          items:
            $ref: "#/components/schemas/UpdateSourceUrlRequest"
    UpdateSourceUrlRequest:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        url:
          type: string
        label:
          type: string
          nullable: true
        browseractWorkflowId:
          type: string
          nullable: true
        active:
          type: boolean
    ScrapeResponse:
      type: object
      properties:
        message:
          type: string
        taskIds:
          type: array
          items:
            type: string
      required: [message]
    ImportEventsRequest:
      type: object
      properties:
        events:
          type: array
          items:
            type: object
      required: [events]
...
    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        introHtml:
          type: string
          nullable: true
        eventsHtml:
          type: string
          nullable: true
        footerHtml:
          type: string
          nullable: true
        isDefault:
          type: boolean
        createdAt:
          type: string
      required: [id, name, isDefault, createdAt]
    CreateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        introHtml:
          type: string
          nullable: true
        eventsHtml:
          type: string
          nullable: true
        footerHtml:
          type: string
          nullable: true
      required: [name]
    UpdateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        introHtml:
          type: string
          nullable: true
        eventsHtml:
          type: string
          nullable: true
        footerHtml:
          type: string
          nullable: true
...
    Settings:
      type: object
      properties:
        publicationDay:
          type: integer
        categories:
          type: array
          items:
            type: string
        cronExpression:
          type: string
        cronEnabled:
          type: boolean
        browseractSinglePageWorkflowId:
          type: string
          nullable: true
        appBaseUrl:
          type: string
          nullable: true
      required: [publicationDay, categories, cronExpression, cronEnabled]
    UpdateSettingsRequest:
      type: object
      properties:
        publicationDay:
          type: integer
        categories:
          type: array
          items:
            type: string
        cronExpression:
          type: string
        cronEnabled:
          type: boolean
        browseractSinglePageWorkflowId:
          type: string
          nullable: true
        appBaseUrl:
          type: string
          nullable: true
        anthropicApiKey:
          type: string
          nullable: true
        browseractApiKey:
          type: string
          nullable: true
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
...