# lib/api-zod/src/generated/types/createTemplateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Varese Life API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTemplateRequest {
  name: string;
  description?: string | null;
  introHtml?: string | null;
  eventsHtml?: string | null;
  footerHtml?: string | null;
}