# lib/api-zod/src/generated/types/updateSourceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Varese Life API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSourceUrlRequest } from "./updateSourceUrlRequest";

export interface UpdateSourceRequest {
  name?: string;
  browseractWorkflowId?: string | null;
  promptOverride?: string | null;
  active?: boolean;
  preferredScraper?: string;
  urls?: UpdateSourceUrlRequest[];
}